requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
aws-cdk-lib==2.131.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
aws-cdk.aws-codestar-alpha==2.131.0a0
billiard==4.2.0
boto3==1.34.54
botocore==1.34.54
cattrs==23.2.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
constructs==10.3.0
cryptography==42.0.5
daphne==4.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environ==1.0
exceptiongroup==1.2.0
hyperlink==21.0.0
idna==3.4
importlib_resources==6.1.2
incremental==22.10.0
iniconfig==2.0.0
jmespath==1.0.1
jsii==1.94.0
kombu==5.3.5
msgpack==1.0.8
packaging==23.2
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
publication==0.0.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytest==8.0.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.2
requests==2.31.0
s3transfer==0.10.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typeguard==2.13.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
